+++ Source/Core/Core/PowerPC/JitInterface.h
@@ -8,6 +8,7 @@
 #include "Common/ChunkFile.h"
 #include "Core/MachineContext.h"
 #include "Core/PowerPC/CPUCoreBase.h"
+#include "Core/PowerPC/Profiler.h"
 
 namespace JitInterface
 {
@@ -25,6 +26,8 @@ namespace JitInterface
 
 	// Debugging
 	void WriteProfileResults(const std::string& filename);
+	void GetProfileResults(ProfileStats* prof_stats);
+	int GetHostCode(u32* address, const u8** code, u32* code_size);
 
 	// Memory Utilities
 	bool HandleFault(uintptr_t access_address, SContext* ctx);
