+++ Source/Core/Core/PowerPC/Profiler.h
@@ -44,12 +44,23 @@
 
 struct BlockStat
 {
-	BlockStat(int bn, u64 c) : blockNum(bn), cost(c) {}
+	BlockStat(int bn, u32 _addr, u64 c, u64 ticks, u32 size) :
+		blockNum(bn), addr(_addr), cost(c), tick_counter(ticks), block_size(size) {}
 	int blockNum;
+	u32 addr;
 	u64 cost;
+	u64 tick_counter;
+	u32 block_size;
 
 	bool operator <(const BlockStat &other) const
 	{ return cost > other.cost; }
+};
+struct ProfileStats
+{
+	std::vector<BlockStat> block_stats;
+	u64 cost_sum;
+	u64 timecost_sum;
+	u64 countsPerSec;
 };
 
 namespace Profiler
