+++ Source/Core/UICommon/UICommon.vcxproj
@@ -44,9 +44,11 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="UICommon.cpp" />
+    <ClCompile Include="Disassembler.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="UICommon.h" />
+    <ClInclude Include="Disassembler.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
